feat: move report entity under user package
